fstrm_io_submit(): no longer do an implicit free on failure
fstrm_io_submit() will now no longer implicitly free the caller's 'buf'
parameter if it wasn't successfully added to the queue. this makes it
possible to implement a "blocking" variant of fstrm_io_submit() that
calls fstrm_io_submit() until it succeeds.

export a trivial "fstrm_free_wrapper" function that just wraps free(),
so the caller doesn't have to do this, in the common case that the 'buf'
parameter is allocated with malloc().

if NULL is passed as the 'free_func' parameter to fstrm_io_submit(),
this now means that no post-processing cleanup should be done on the
caller's buffer, at all.