fstrm: add 'queue_model' option
this option will allow selecting between the SPSC (memory barrier) and
MPSC (mutex) queue implementations if both are available. in the worst
case, the memory barrier implementation isn't available on this platform
and we have to fall back to the mutex implementation (which will always
be available), but that's OK.

this requires an fstrm_queue_model enum and an
fstrm_io_options_set_queue_model() function.