doxygen/Doxyfile.in: add our changes
note that the initial commit for this file contains the verbatim
"doxygen -g" output from doxygen 1.8.6, so by implementing our changes
in a subsequent commit we get a precise record of what we changed from
the default template doxygen config file.

also note that any project-specific settings get substituted in via
autoconf (doxygen/Doxyfile is listed in AC_CONFIG_FILES), so this file
can be reused verbatim in other projects.