fstrm_capture: new
This is a reference "capture utility". It uses libevent to multiplex
incoming Frame Streams payloads on a local socket into a single output
file. It requires specifying the desired content type on the command
line. Example command-line use:

  $ fstrm_capture -t protobuf:dnstap.Dnstap -u /tmp/dnstap.sock -w /tmp/dnstap.out

It currently has a hardcoded 256K limit on the size of inbound frames,
in order to bound the amount of memory used to buffer an individual
frame on a particular connection. This needs to be fixed so that the
limit can be configured, and so that frames which exceed this limit are
properly drained and silently skipped. (Another alternative is to simply
drop the connection.)

Another thing that needs to be fixed is to limit the number of open
connections from Frame Streams clients, and to make this limit
configurable.