tcp_writer: New writer implementation that connects to a TCP socket
This commit adds 'tcp_writer' to the core library, which implements a
bi-directional Frame Streams writer as a TCP socket client.

The rdwr operation methods are copied with minor changes from
unix_writer. It may be worthwhile to factor out the "base" functionality
into generic socket methods, but I've not attempted to do so in this
commit.