fstrm_capture: complete sending of FINISH frame before closing connection.
The FINISH control frame was being committed to the write buffer, but
the connection was closed immediately thereafter, meaning libevent was
unable to schedule a write to the connection. We solve this by closing
the connection from a write callback, which is called when the control
frame has been written to the connection.