Rename options and rewrite descriptions for brevity.
The argv.[ch] argument processing system in libmy
truncates overly long option names and descriptions
for help formatting purposes. This rewrite makes those
names and descriptions fit those length limitations.

This also fixes an error in the description of the
connections / maxconns option: the default behavior
is to have no limit, not 128.